requirements.txt
firebase-admin==6.0.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
flask-msearch==0.2.5
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
Jinja2==2.11.1
keras==2.9.0
numpy==1.23.5
Pyrebase4==4.5.0
SQLAlchemy==1.3.13
WTForms==2.2.1
tensorflow==2.9.2
transformers==4.25.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
werkzeug==2.0.3